requirements.txt
dvc
dvc_gdrive
numpy
pandas
telethon
python-dotenv